Drop advanced from logic (#29)
Dropping advanced form generation and display. This functionality provides more complexity than the feature requires.

The feature was meant to allow customizing the form, but the average user who requires form customization doesn't know anything about customizing HTML. The tool is too advanced for non-developers and doesn't provide any extra value for a developer. The functionality can be implemented using the built-in functionality of a custom HTML block.

* Remove all code regarding advanced form functionality