Drop templater class and unify API queries (#34)
The custom template class went against WordPress's "escape late" logic. The template was built so that the values provided to it should be escaped. Escaping should happen late and in the exact place, the variable is echoed. Also, the Smaily_Request class needs to be refactored to improve the API and expose available actions.

* Initialize Smaily_Request with credentials on initialization
* Expose method instead of raw queries from Smaily_Request
* Delete smaily-template class
* Refactor widget and shortcode rendering logic