Drop templater class and unify API queries (#35)
The custom template class went against WordPress's "escape late" logic. The template was built so that the values provided to it should be escaped. Escaping should happen late when the variable is echoed. Also, the Smaily_Request class needs to be refactored to improve the API and expose available actions.


* Drop the Smaily_Template class 
* Refactor Smaily_Requests class
* Move message rendering to the basic template