Fix: Fix profile settings translations not loaded
Fixes an issue where translations were not loaded for profile settings additional fields. This was caused because the construct method runs during 'init' hook where translations are not available. Loading the field values when rendering the fields for the first time fixes the issue.

Also I improved the styling of the gender fields by putting them beneath one another.