blocks/newsletter-form: Replace Card component with div
Replacing the Card component with a regular div. There seems to be an error where the child container height is automatically 100% which may cause issues when rendering the Smaily Opt In block without parent container block such as row/group etc. The regular div seems to behave more consistently in every occasion.