sandbox: Update sandbox dependencies, define minimum supported version config
Adds getttext and zip tools to sandbox to allow extracting translations and package the plugin inside the container.

Adds a separate composer file to store minimum supported version configuration. This allows developer to easily switch between WP versions and check if the module works also with the minimum supported version.
Also this makes it clear which functions are available to the developer when we pump up the minimum supported version of WordPress.