Fix: Remove random ordering of RSS feed items (#94)
Removes random ordering of RSS feed items. This is a broken option as the subsequent requests to the same URL do not give out the same feed items. This causes feed item loading errors when the RSS-importer tries to load RSS-feed items. The random ordering and limit combination may cause selected items not to appear.

Random ordering is also not valid from logical perspective as the user has to pick the items when using the RSS-feed block in the editor. The email would not receive random items anyway so the option is also logically invalid.