integrations/elementor: Reduce number of API calls (#102)
The Elementor widget frontend makes an API request to load autoresponders. This feature is only needed in the admin view, where the user can select an autoresponder for the form. The selected autoresponder ID is saved and used when rendering the frontend widget. This PR improves the performance of the Elementor widget by allowing autoresponder fetching only in the admin view.